First Mac Commit
-Now supports Mac OSX
-Fixed Finger->Hand crash on exit
-Update to Leap SDK 2.2.1
