--- Plugins/LeapMotion/Source/LeapMotion/Private/AnimBody/CreateNewObject.cpp
@@ -1,19 +1,0 @@
-//Copyleft under the creative commons license
-//For details see http://creativecommons.org/licenses/by-sa/4.0/
-
-#include "LeapMotionPrivatePCH.h"
-#include "CreateNewObject.h"
-
-UCreateNewObject::UCreateNewObject(const class FObjectInitializer& PCIP)
-	: Super(PCIP)
-{
-
-}
-
-UObject* UCreateNewObject::NewObjectFromBlueprint(UObject* WorldContextObject, TSubclassOf<UObject> UC)
-{
-	UWorld* World = GEngine->GetWorldFromContextObject(WorldContextObject);
-	UObject* tempObject = StaticConstructObject(UC);
-
-	return tempObject;
-}
