+++ Plugins/LeapMotion/Source/LeapMotion/Private/LeapFinger.cpp
@@ -44,6 +44,8 @@ ULeapFinger::~ULeapFinger()
 
 void ULeapFinger::CleanupRootReferences()
 {
+    ULeapPointable::CleanupRootReferences();
+    
 	if (this->HasAnyFlags(RF_RootSet))
 		this->RemoveFromRoot();
 }
